Changed pin variable types

+++ 7segment-clock.ino
@@ -7,18 +7,18 @@
 const int button1_pin = 4;
 const int button2_pin = 3;
 
-const int segA = 8;
-const int segB = 9;
-const int segC = 10;
-const int segD = 11;
-const int segE = 12;
-const int segF = 13;
-const int segG = 7;
+const byte segA = 8;
+const byte segB = 9;
+const byte segC = 10;
+const byte segD = 11;
+const byte segE = 12;
+const byte segF = 13;
+const byte segG = 7;
 
-const int anode1 = A3;
-const int anode2 = A2;
-const int anode3 = A1;
-const int anode4 = A0;
+const byte anode1 = A3;
+const byte anode2 = A2;
+const byte anode3 = A1;
+const byte anode4 = A0;
 
 bool button1_last_pressed = false;
 bool button2_last_pressed = false;
